# backend/openapi_agent.yaml
openapi: 3.0.3
info:
  title: CEMA Agent Service API
  description: Microserviço de Agente do Sistema CEMA (Google ADK).
  version: 1.0.0
  contact:
    name: CEMA Team

servers:
  - url: http://localhost:8080
    description: Agent Service

tags:
  - name: Agent
    description: Execução do agente
  - name: Sessions
    description: Gerenciamento de sessões
  - name: Health
    description: Verificação de status

paths:
  /api/health:
    get:
      summary: Health check do Agent Service
      tags:
        - Health
      responses:
        '200':
          description: Agent Service funcionando
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: agent-service

  /api/apps/{appName}/users/{userId}/sessions/{sessionId}:
    post:
      summary: Criar nova sessão
      tags:
        - Sessions
      security:
        - bearerAuth: []
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
          example: cema_system
        - name: userId
          in: path
          required: true
          schema:
            type: string
          example: user_123
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          example: session_456
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              initialContext: "Análise de mercado tech"
      responses:
        '200':
          description: Sessão criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      summary: Obter detalhes da sessão
      tags:
        - Sessions
      security:
        - bearerAuth: []
      parameters:
        - name: appName
          in: path
          required: true
          schema:
            type: string
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dados da sessão
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '404':
          description: Sessão não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /api/run:
    post:
      summary: Executar agente
      description: Envia mensagem para o agente e retorna todos os eventos
      tags:
        - Agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunRequest'
            example:
              appName: cema_system
              userId: user_123
              sessionId: session_456
              newMessage:
                role: user
                parts:
                  - text: "Analise as tendências do setor de tecnologia"
      responses:
        '200':
          description: Eventos do agente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '400':
          description: Campos obrigatórios faltando
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Não autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Session:
      type: object
      properties:
        id:
          type: string
          example: session_456
        appName:
          type: string
          example: cema_system
        userId:
          type: string
          example: user_123
        state:
          type: object
        events:
          type: array
          items:
            type: object
        lastUpdateTime:
          type: string
          format: date-time

    RunRequest:
      type: object
      required:
        - appName
        - userId
        - sessionId
        - newMessage
      properties:
        appName:
          type: string
        userId:
          type: string
        sessionId:
          type: string
        newMessage:
          type: object
          properties:
            role:
              type: string
              enum: [user, model]
            parts:
              type: array
              items:
                type: object
                properties:
                  text:
                    type: string

    Event:
      type: object
      properties:
        author:
          type: string
          example: cso_agent
        content:
          type: object
          properties:
            role:
              type: string
            parts:
              type: array
              items:
                type: object

    Error:
      type: object
      properties:
        detail:
          type: string
          example: Error message

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token necessário.